requirements.txt
aiofiles==23.2.1
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.9.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
cryptography==43.0.1
distlib==0.3.9
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fake-useragent==0.1.11
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
googletrans==3.1.0a0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.15.0
httpx==0.23.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
libmesh==0.0.6
magic-filter==1.0.12
MarkupSafe==2.1.5
MouseInfo==0.1.3
mss==10.0.0
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.2.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
psutil==6.1.0
pyaes==1.6.1
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pynput==1.7.7
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-xlib==0.33
pytweening==1.2.0
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
runware==0.3.2
scapy==2.5.0
schedule==1.2.2
secure-smtplib==0.1.1
six==1.17.0
sniffio==1.3.1
tbomb==2.1.2
Telethon==1.38.1
thread6==0.2.0
typing_extensions==4.12.2
urllib3==1.26.6
virtualenv==20.27.1
websockets==12.0
Werkzeug==3.0.3
wget==3.2
yarl==1.17.0
